last

+++ libft/submit/ft_strlcat.c
@@ -6,13 +6,13 @@
 /*   By: nugoncal <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/16 18:00:30 by nugoncal          #+#    #+#             */
-/*   Updated: 2025/04/17 10:56:08 by nugoncal         ###   ########.fr       */
+/*   Updated: 2025/05/08 16:43:26 by nugoncal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcat(char *dest, const char *src, unsigned int destsize)
+size_t	ft_strlcat(char *dest, const char *src, unsigned int destsize)
 {
 	unsigned int	i;
 	unsigned int	dest_len;
